weather_process/requirements.txt
fastapi
joblib
kafka-python
kafka-python
pandas
psycopg2-binary
python-dotenv
scikit-learn